feat(context): Added useContext for data to be available throughout the app
