+++ src/App.jsx
@@ -1,11 +1,15 @@
 import Login from "./components/Login"
 import Main from "./components/Main-window"
+import {useContext} from 'react';
+import {Context} from "./Context"
 import './App.css'
 
 function App() {
+  const {accessToken} = useContext(Context);
+
   return (
     <>
-      {false ? <Main /> : <Login />}
+      {accessToken ? <Main /> : <Login />}
 
     </>
   )
